Asana 1147789120513007 - TBPP - Eviter que certaines options s'enregistrent quand elles ne s'affichent pas (non-admin)

+++ includes/control/ajax.php
@@ -759,34 +759,34 @@ class WDGAjaxActions {
 			}
 		}
 		
-		// Masquer au public
-		$new_is_hidden = filter_input( INPUT_POST, 'new_is_hidden');
-        if ( $new_is_hidden === true || $new_is_hidden === "true" || $new_is_hidden === 1 ) {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_campaign_is_hidden, '1' );
-		} else {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_campaign_is_hidden );
-		}
-		$success[ 'new_is_hidden' ] = 1;
-		
-		// Passer la phase de vote
-		$new_skip_vote = filter_input( INPUT_POST, 'new_skip_vote');
-        if ( $new_skip_vote === true || $new_skip_vote === "true" || $new_skip_vote === 1 ) {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_skip_vote, '1' );
-		} else {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_skip_vote );
-		}
-		$success[ 'new_skip_vote' ] = 1;
-		
-		// Ne pas compter dans les stats
-		$new_skip_in_stats = filter_input( INPUT_POST, 'new_skip_in_stats' );
-        if ( $new_skip_in_stats === true || $new_skip_in_stats === "true" || $new_skip_in_stats === 1 ) {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_skip_in_stats, '1' );
-		} else {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_skip_in_stats );
-		}
-		$success[ 'new_skip_in_stats' ] = 1;
-
 		if ( $current_wdg_user->is_admin() ) {
+			// Masquer au public
+			$new_is_hidden = filter_input( INPUT_POST, 'new_is_hidden');
+			if ( $new_is_hidden === true || $new_is_hidden === "true" || $new_is_hidden === 1 ) {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_campaign_is_hidden, '1' );
+			} else {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_campaign_is_hidden );
+			}
+			$success[ 'new_is_hidden' ] = 1;
+			
+			// Passer la phase de vote
+			$new_skip_vote = filter_input( INPUT_POST, 'new_skip_vote');
+			if ( $new_skip_vote === true || $new_skip_vote === "true" || $new_skip_vote === 1 ) {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_skip_vote, '1' );
+			} else {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_skip_vote );
+			}
+			$success[ 'new_skip_vote' ] = 1;
+			
+			// Ne pas compter dans les stats
+			$new_skip_in_stats = filter_input( INPUT_POST, 'new_skip_in_stats' );
+			if ( $new_skip_in_stats === true || $new_skip_in_stats === "true" || $new_skip_in_stats === 1 ) {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_skip_in_stats, '1' );
+			} else {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_skip_in_stats );
+			}
+			$success[ 'new_skip_in_stats' ] = 1;
+
 			//Catégories du projet
 			$new_project_categories = array();
 			if ( isset( $_POST["new_project_categories"] ) ) $new_project_categories = $_POST["new_project_categories"];
@@ -838,41 +838,44 @@ class WDGAjaxActions {
 			$success[ "new_minimum_goal_display" ] = 1;
 		}
 		
-		$new_enable_advice_notifications = sanitize_text_field( filter_input( INPUT_POST, 'new_enable_advice_notifications' ) );
-		$queued_action_id = $campaign->has_planned_advice_notification();
-        if ( $new_enable_advice_notifications === true || $new_enable_advice_notifications === "true" || $new_enable_advice_notifications === 1 ) {
-			if ( $queued_action_id == FALSE ) {
-				WDGQueue::add_campaign_advice_notification( $campaign->ID );
+		if ( $current_wdg_user->is_admin() ) {
+			$new_enable_advice_notifications = sanitize_text_field( filter_input( INPUT_POST, 'new_enable_advice_notifications' ) );
+			$queued_action_id = $campaign->has_planned_advice_notification();
+			if ( $new_enable_advice_notifications === true || $new_enable_advice_notifications === "true" || $new_enable_advice_notifications === 1 ) {
+				if ( $queued_action_id == FALSE ) {
+					WDGQueue::add_campaign_advice_notification( $campaign->ID );
+				}
+			} else {
+				if ( $queued_action_id != FALSE ) {
+					WDGWPREST_Entity_QueuedAction::edit( $queued_action_id, WDGQueue::$status_complete );
+				}
 			}
-		} else {
-			if ( $queued_action_id != FALSE ) {
-				WDGWPREST_Entity_QueuedAction::edit( $queued_action_id, WDGQueue::$status_complete );
+		
+			$new_show_comments_for_everyone = sanitize_text_field( filter_input( INPUT_POST, 'new_show_comments_for_everyone' ) );
+			if ( $new_show_comments_for_everyone === true || $new_show_comments_for_everyone === "true" || $new_show_comments_for_everyone === 1 ) {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_show_comments_for_everyone, '1' );
+			} else {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_show_comments_for_everyone );
+			}
+		
+			$new_hide_investors = sanitize_text_field( filter_input( INPUT_POST, 'new_hide_investors' ) );
+			if ( $new_hide_investors === true || $new_hide_investors === "true" || $new_hide_investors === 1 ) {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_hide_investors, '1' );
+			} else {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_hide_investors );
+			}
+			$new_can_invest_until_contract_start_date = sanitize_text_field( filter_input( INPUT_POST, 'new_can_invest_until_contract_start_date' ) );
+			if ( $new_can_invest_until_contract_start_date === true || $new_can_invest_until_contract_start_date === "true" || $new_can_invest_until_contract_start_date === 1 ) {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_can_invest_until_contract_start_date, '1' );
+			} else {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_can_invest_until_contract_start_date );
 			}
 		}
-		
-		$new_show_comments_for_everyone = sanitize_text_field( filter_input( INPUT_POST, 'new_show_comments_for_everyone' ) );
-        if ( $new_show_comments_for_everyone === true || $new_show_comments_for_everyone === "true" || $new_show_comments_for_everyone === 1 ) {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_show_comments_for_everyone, '1' );
-		} else {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_show_comments_for_everyone );
-		}
-		
-		$new_hide_investors = sanitize_text_field( filter_input( INPUT_POST, 'new_hide_investors' ) );
-        if ( $new_hide_investors === true || $new_hide_investors === "true" || $new_hide_investors === 1 ) {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_hide_investors, '1' );
-		} else {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_hide_investors );
-		}
+
 		$new_archive_message = sanitize_text_field( filter_input( INPUT_POST, 'new_archive_message' ) );
 		if ( !empty( $new_archive_message ) ) {
 			$campaign->__set( ATCF_Campaign::$key_archive_message, $new_archive_message );
 			$success[ "new_archive_message" ] = 1;
-		}
-		$new_can_invest_until_contract_start_date = sanitize_text_field( filter_input( INPUT_POST, 'new_can_invest_until_contract_start_date' ) );
-        if ( $new_can_invest_until_contract_start_date === true || $new_can_invest_until_contract_start_date === "true" || $new_can_invest_until_contract_start_date === 1 ) {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_can_invest_until_contract_start_date, '1' );
-		} else {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_can_invest_until_contract_start_date );
 		}
 		$new_end_vote_pending_message = sanitize_text_field( filter_input( INPUT_POST, 'new_end_vote_pending_message' ) );
 		if ( !empty( $new_end_vote_pending_message ) ) {
@@ -889,20 +892,23 @@ class WDGAjaxActions {
 			$campaign->__set( ATCF_Campaign::$key_custom_footer_code, $new_custom_footer_code );
 			$success[ "new_custom_footer_code" ] = 1;
 		}
-		$new_is_check_payment_available = filter_input( INPUT_POST, 'new_is_check_payment_available');
-        if ( $new_is_check_payment_available === true || $new_is_check_payment_available === "true" || $new_is_check_payment_available === 1 ) {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_can_use_check );
-		} else {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_can_use_check, '0' );
+		
+		if ( $current_wdg_user->is_admin() ) {
+			$new_is_check_payment_available = filter_input( INPUT_POST, 'new_is_check_payment_available');
+			if ( $new_is_check_payment_available === true || $new_is_check_payment_available === "true" || $new_is_check_payment_available === 1 ) {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_can_use_check );
+			} else {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_can_use_check, '0' );
+			}
+			$success[ 'new_is_check_payment_available' ] = 1;
+			$new_has_overridden_wire_constraints = filter_input( INPUT_POST, 'new_has_overridden_wire_constraints');
+			if ( $new_has_overridden_wire_constraints === true || $new_has_overridden_wire_constraints === "true" || $new_has_overridden_wire_constraints === 1 ) {
+				update_post_meta( $campaign_id, ATCF_Campaign::$key_has_overridden_wire_constraints, '1' );
+			} else {
+				delete_post_meta( $campaign_id, ATCF_Campaign::$key_has_overridden_wire_constraints );
+			}
+			$success[ 'new_has_overridden_wire_constraints' ] = 1;
 		}
-		$success[ 'new_is_check_payment_available' ] = 1;
-		$new_has_overridden_wire_constraints = filter_input( INPUT_POST, 'new_has_overridden_wire_constraints');
-        if ( $new_has_overridden_wire_constraints === true || $new_has_overridden_wire_constraints === "true" || $new_has_overridden_wire_constraints === 1 ) {
-			update_post_meta( $campaign_id, ATCF_Campaign::$key_has_overridden_wire_constraints, '1' );
-		} else {
-			delete_post_meta( $campaign_id, ATCF_Campaign::$key_has_overridden_wire_constraints );
-		}
-		$success[ 'new_has_overridden_wire_constraints' ] = 1;
 		
 		$new_fake_url = filter_input( INPUT_POST, 'new_fake_url' );
 		if ( !empty( $new_fake_url ) ) {
